+++ src/components/api-address-form/index.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react'
 import { connect } from 'redux-bundler-react'
 import { withTranslation } from 'react-i18next'
-import Button from '../button/button.tsx'
+import Button from '../button/button'
 import { checkValidAPIAddress } from '../../bundles/ipfs-provider.js'
 
 const ApiAddressForm = ({ t, doUpdateIpfsApiAddress, ipfsApiAddress, ipfsInitFailed }) => {
