+++ src/components/is-not-connected/IsNotConnected.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react'
 import { connect } from 'redux-bundler-react'
 import { withTranslation, Trans } from 'react-i18next'
 import classNames from 'classnames'
-import ApiAddressForm from '../api-address-form/ApiAddressForm.js'
+import ApiAddressForm from '../api-address-form/index'
 import Box from '../box/Box.js'
 import Shell from '../shell/Shell.js'
 import GlyphAttention from '../../icons/GlyphAttention.js'
